requirements.txt
openai
langchain
langchain-openai
langchainhub
chromadb
python-dotenv
tiktoken
streamlit
Pillow
langchain-community
langchain_chroma
pinecone
